meroctl shouldn't need to depend on calimero-server
