+++ crates/server-primitives/Cargo.toml
@@ -14,6 +14,7 @@ serde_json.workspace = true
 thiserror.workspace = true
 url = { workspace = true, features = ["serde"] }
 
+calimero-context-config.workspace = true
 calimero-node-primitives.workspace = true
 calimero-primitives.workspace = true
